refactor(app): 拆出全局右键菜单规范化与定位纯函数
- 新增 utils/globalContextMenu.ts：菜单项规范化、位置解析、可编辑目标判定（纯函数与共享类型）
- GlobalContextMenu.tsx 416 -> 338 行
