+++ frontend/src/components/GlobalContextMenu.tsx
@@ -2,97 +2,19 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { useTranslation } from '../i18n.ts';
 import { formatShortcut } from '../utils/platform.ts';
-import { clampMenuPosition } from '../utils/menuPosition.ts';
 import { GLOBAL_CONTEXT_MENU_OPEN_EVENT, type GlobalContextMenuDetail } from '../utils/contextMenu.ts';
+import {
+  type ContextMenuItemInput,
+  type ContextMenuPositionDetail,
+  type EditableAction,
+  type NormalizedMenuItem,
+  normalizeMenuItems,
+  resolveEditableTarget,
+  resolveMenuPosition,
+} from '../utils/globalContextMenu.ts';
 import { MenuList, MenuPanel, type MenuItem } from './ui';
 import { Z } from '../constants/zIndex.ts';
 import * as runtime from '../../wailsjs/runtime/runtime.js';
-
-/** 菜单项输入（来自 .tsx 调用方，字段宽松） */
-interface ContextMenuItemInput {
-  type?: string;
-  key?: string;
-  label?: string;
-  shortcut?: string;
-  danger?: boolean;
-  disabled?: boolean;
-  children?: unknown;
-  onSelect?: unknown;
-}
-
-type NormalizedMenuItem =
-  | { key: string; type: 'divider' }
-  | {
-      key: string;
-      type: 'item';
-      label: string;
-      shortcut: string;
-      danger: boolean;
-      disabled: boolean;
-      children: NormalizedMenuItem[];
-      onSelect: ((item: NormalizedMenuItem) => void) | null;
-    };
-
-function normalizeMenuItems(items: unknown): NormalizedMenuItem[] {
-  if (!Array.isArray(items)) {
-    return [];
-  }
-  return items
-    .map((item: ContextMenuItemInput, index: number): NormalizedMenuItem | null => {
-      if (item?.type === 'divider') {
-        return {
-          key: typeof item?.key === 'string' && item.key.trim() ? item.key.trim() : `divider-${index}`,
-          type: 'divider',
-        };
-      }
-      const label = typeof item?.label === 'string' ? item.label.trim() : '';
-      if (!label) {
-        return null;
-      }
-      const onSelect = item?.onSelect;
-      const rawChildren = item?.children;
-      return {
-        key: typeof item?.key === 'string' && item.key.trim() ? item.key.trim() : `item-${index}`,
-        type: 'item',
-        label,
-        shortcut: typeof item?.shortcut === 'string' ? item.shortcut.trim() : '',
-        danger: item?.danger === true,
-        disabled: item?.disabled === true,
-        children: Array.isArray(rawChildren) ? normalizeMenuItems(rawChildren) : [],
-        onSelect: typeof onSelect === 'function' ? (onSelect as (item: NormalizedMenuItem) => void) : null,
-      };
-    })
-    .filter((item): item is NormalizedMenuItem => item !== null);
-}
-
-interface ContextMenuPositionDetail {
-  x?: unknown;
-  y?: unknown;
-  estimatedWidth?: unknown;
-  estimatedHeight?: unknown;
-}
-
-function resolveMenuPosition(detail: ContextMenuPositionDetail, itemCount: number) {
-  const x = Number(detail?.x);
-  const y = Number(detail?.y);
-  const estimatedWidth = Number(detail?.estimatedWidth);
-  const estimatedHeight = Number(detail?.estimatedHeight);
-  const width = Number.isFinite(estimatedWidth) && estimatedWidth > 0 ? estimatedWidth : 168;
-  const height = Number.isFinite(estimatedHeight) && estimatedHeight > 0 ? estimatedHeight : Math.max(40, itemCount * 36 + 8);
-  return clampMenuPosition(Number.isFinite(x) ? x : 0, Number.isFinite(y) ? y : 0, width, height);
-}
-
-function resolveEditableTarget(target: EventTarget | null): HTMLInputElement | HTMLTextAreaElement | null {
-  if (typeof window === 'undefined') {
-    return null;
-  }
-  if (target instanceof window.HTMLInputElement || target instanceof window.HTMLTextAreaElement) {
-    return target;
-  }
-  return null;
-}
-
-type EditableAction = 'copy' | 'cut' | 'paste' | 'selectAll';
 
 interface MenuState {
   x: number;
